+++ CHANGELOG.rst
@@ -6,6 +6,18 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog`_,
 and this project adheres to `Semantic Versioning`_.
 
+`1.3.3`_ -  2020-08-31
+--------------------------
+Added
+'''''
+- Helper functions for SeedList class.
+
+Fixed
+'''''''
+- Dictionary conversion issue with lists of SciPy distributions.
+- XML tags in documentation on position distributions.
+
+
 `1.3.2`_ -  2020-07-11
 --------------------------
 Added
@@ -129,7 +141,8 @@ Added
 
 .. LINKS
 
-.. _`Unreleased`: https://github.com/kip-hart/MicroStructPy/compare/v1.3.2...HEAD
+.. _`Unreleased`: https://github.com/kip-hart/MicroStructPy/compare/v1.3.3...HEAD
+.. _`1.3.3`: https://github.com/kip-hart/MicroStructPy/compare/v1.3.2...v1.3.3
 .. _`1.3.2`: https://github.com/kip-hart/MicroStructPy/compare/v1.3.1...v1.3.2
 .. _`1.3.1`: https://github.com/kip-hart/MicroStructPy/compare/v1.3.0...v1.3.1
 .. _`1.3.0`: https://github.com/kip-hart/MicroStructPy/compare/v1.2.2...v1.3.0